Fix f32 precision loss in zoom-to-event padding calculation
Use f64 instead of f32 for the 20% duration padding calculation.
f32 only has ~7 digits of mantissa precision, losing accuracy for
durations above 16ms (16_000_000 ns) which are common in profiling.

+++ src/main.rs
@@ -439,7 +439,7 @@ impl Lineme {
                     let event_rel_end = event_rel_start.saturating_add(event.duration_ns);
 
                     // Add padding of 20% of event duration (10% on each side)
-                    let padding_ns = ((event.duration_ns as f32) * 0.2).round() as u64;
+                    let padding_ns = ((event.duration_ns as f64) * 0.2).round() as u64;
                     let half_pad = padding_ns / 2;
 
                     let start_ns = event_rel_start.saturating_sub(half_pad).min(total_ns);
